# apps/backend/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login with site code, email and password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Refresh access token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/oauth2/{provider}/authorize": {
      "get": {
        "operationId": "AuthController_authorizeOAuth2",
        "summary": "Initiate OAuth2/OIDC login flow",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/oauth2/{provider}/callback": {
      "get": {
        "operationId": "AuthController_oauth2Callback",
        "summary": "OAuth2/OIDC callback handler",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout and revoke refresh token",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "summary": "Request password reset - sends OTP to email",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "Company ID (UUID) for the tenant. This should be the company UUID resolved from the public lookup / company selection flow.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "summary": "Verify OTP and get reset token",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "Company ID (UUID) for the tenant. This should be the company UUID resolved from the public lookup / company selection flow.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "summary": "Reset password using token from OTP verification",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "Company ID (UUID) for the tenant. This should be the company UUID resolved from the public lookup / company selection flow.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/dashboard/stats": {
      "get": {
        "operationId": "DashboardController_getStats",
        "summary": "Get dashboard statistics",
        "description": "Returns aggregated statistics for the dashboard including service request counts, user counts, and completion metrics.",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Optional role filter for role-specific statistics",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStatsDto"
                }
              }
            }
          }
        },
        "tags": [
          "dashboard"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/dashboard/analytics": {
      "get": {
        "operationId": "DashboardController_getAnalytics",
        "summary": "Get dashboard analytics",
        "description": "Returns detailed analytics including distributions, trends, performance metrics, and predictions.",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Optional role filter for role-specific analytics",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "description": "Time period for trend analysis",
            "schema": {
              "enum": [
                "daily",
                "weekly",
                "monthly"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard analytics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardAnalyticsDto"
                }
              }
            }
          }
        },
        "tags": [
          "dashboard"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TENX API",
    "description": "Multi-tenant SaaS ERP backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "siteCode": {
            "type": "string",
            "example": "WIPRP-HQ",
            "description": "Site code for tenant scoping"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "companyId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "siteCode",
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "otp": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "description": "New password (minimum 8 characters)"
          }
        },
        "required": [
          "newPassword"
        ]
      },
...
      "DashboardStatsDto": {
        "type": "object",
        "properties": {
          "openRequests": {
            "type": "number",
            "description": "Number of open tickets"
          },
          "inProgress": {
            "type": "number",
            "description": "Number of tickets in progress"
          },
          "resolved": {
            "type": "number",
            "description": "Number of resolved tickets"
          },
          "totalRequests": {
            "type": "number",
            "description": "Total number of tickets"
          },
          "activeUsers": {
            "type": "number",
            "description": "Number of active users"
          },
          "pendingApproval": {
            "type": "number",
            "description": "Number of pending approvals"
          },
          "completedToday": {
            "type": "number",
            "description": "Number of tickets completed today"
          },
          "assigned": {
            "type": "number",
            "description": "Number of assigned tickets"
          },
          "completed": {
            "type": "number",
            "description": "Number of completed tickets"
          },
          "pending": {
            "type": "number",
            "description": "Number of pending tickets"
          },
          "acknowledged": {
            "type": "number",
            "description": "Number of acknowledged tickets"
          },
          "onHold": {
            "type": "number",
            "description": "Number of tickets on hold"
          },
          "escalated": {
            "type": "number",
            "description": "Number of escalated tickets"
          },
          "overdue": {
            "type": "number",
            "description": "Number of overdue tickets"
          }
        },
        "required": [
          "openRequests",
          "inProgress",
          "resolved",
          "totalRequests",
          "activeUsers",
          "pendingApproval",
          "completedToday",
          "assigned",
          "completed",
          "pending",
          "acknowledged",
          "onHold"
        ]
      },
      "StatusDistributionDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status name"
          },
          "count": {
            "type": "number",
            "description": "Count of tickets with this status"
          },
          "percentage": {
            "type": "number",
            "description": "Percentage of total tickets"
          }
        },
        "required": [
          "status",
          "count",
          "percentage"
        ]
      },
      "PriorityDistributionDto": {
        "type": "object",
        "properties": {
          "priority": {
            "type": "string",
            "description": "Priority level"
          },
          "count": {
            "type": "number",
            "description": "Count of tickets with this priority"
          },
          "percentage": {
            "type": "number",
            "description": "Percentage of total tickets"
          }
        },
        "required": [
          "priority",
          "count",
          "percentage"
        ]
      },
      "DepartmentDistributionDto": {
        "type": "object",
        "properties": {
          "departmentName": {
            "type": "string",
            "description": "Department name"
          },
          "count": {
            "type": "number",
            "description": "Count of tickets for this department"
          },
          "percentage": {
            "type": "number",
            "description": "Percentage of total tickets"
          }
        },
        "required": [
          "departmentName",
          "count",
          "percentage"
        ]
      },
      "VillaDistributionDto": {
        "type": "object",
        "properties": {
          "villaNumber": {
            "type": "string",
            "description": "Villa number"
          },
          "count": {
            "type": "number",
            "description": "Count of tickets for this villa"
          },
          "percentage": {
            "type": "number",
            "description": "Percentage of total tickets"
          }
        },
        "required": [
          "villaNumber",
          "count",
          "percentage"
        ]
      },
      "TechnicianPerformanceDto": {
        "type": "object",
        "properties": {
          "technicianId": {
            "type": "string",
            "description": "Technician ID"
          },
          "technicianName": {
            "type": "string",
            "description": "Technician name"
          },
          "totalTickets": {
            "type": "number",
            "description": "Total tickets assigned"
          },
          "completedTickets": {
            "type": "number",
            "description": "Tickets completed"
          },
          "averageResolutionTime": {
            "type": "number",
            "description": "Average resolution time in hours"
          },
          "completionRate": {
            "type": "number",
            "description": "Completion rate percentage"
          }
        },
        "required": [
          "technicianId",
          "technicianName",
          "totalTickets",
          "completedTickets",
          "averageResolutionTime",
          "completionRate"
        ]
      },
      "TrendDataPointDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Date in ISO format"
          },
          "created": {
            "type": "number",
            "description": "Number of tickets created"
          },
          "resolved": {
            "type": "number",
            "description": "Number of tickets resolved"
          },
          "inProgress": {
            "type": "number",
            "description": "Number of tickets in progress"
          }
        },
        "required": [
          "date",
          "created",
          "resolved",
          "inProgress"
        ]
      },
      "PerformanceMetricsDto": {
        "type": "object",
        "properties": {
          "averageResolutionTime": {
            "type": "number",
            "description": "Average resolution time in hours"
          },
          "averageFirstResponseTime": {
            "type": "number",
            "description": "Average first response time in hours"
          },
          "slaCompliance": {
            "type": "number",
            "description": "SLA compliance percentage"
          },
          "ticketsResolved": {
            "type": "number",
            "description": "Total tickets resolved this period"
          }
        },
        "required": [
          "averageResolutionTime",
          "averageFirstResponseTime",
          "slaCompliance",
          "ticketsResolved"
        ]
      },
      "PredictionDto": {
        "type": "object",
        "properties": {
          "predictedVolume": {
            "type": "number",
            "description": "Predicted ticket volume for next period"
          },
          "confidence": {
            "type": "number",
            "description": "Confidence level (0-100)"
          },
          "trend": {
            "type": "string",
            "description": "Trend direction (increasing, decreasing, stable)"
          },
          "expectedResolutionTime": {
            "type": "number",
            "description": "Expected resolution time in hours"
          }
        },
        "required": [
          "predictedVolume",
          "confidence",
          "trend",
          "expectedResolutionTime"
        ]
      },
      "DashboardAnalyticsDto": {
        "type": "object",
        "properties": {
          "statusDistribution": {
            "description": "Status distribution",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatusDistributionDto"
            }
          },
          "priorityDistribution": {
            "description": "Priority distribution",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriorityDistributionDto"
            }
          },
          "departmentDistribution": {
            "description": "Department distribution",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DepartmentDistributionDto"
            }
          },
          "villaDistribution": {
            "description": "Villa distribution",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VillaDistributionDto"
            }
          },
          "technicianPerformance": {
            "description": "Technician performance",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TechnicianPerformanceDto"
            }
          },
          "trends": {
            "description": "Trend data over time",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrendDataPointDto"
            }
          },
          "performance": {
            "description": "Performance metrics",
            "allOf": [
              {
                "$ref": "#/components/schemas/PerformanceMetricsDto"
              }
            ]
          },
          "predictions": {
            "description": "Predictions",
            "allOf": [
              {
                "$ref": "#/components/schemas/PredictionDto"
              }
            ]
          }
        },
        "required": [
          "statusDistribution",
          "priorityDistribution",
          "departmentDistribution",
          "villaDistribution",
          "technicianPerformance",
          "trends",
          "performance",
          "predictions"
        ]
      },
...
    }
  }
}